+++ frontend/src/redux/ducks/user/user.slice.ts
@@ -20,7 +20,7 @@ const initialState: UserState = {
 };
 
 export const userSlice = createSlice({
-	name: "users",
+	name: "user",
 	initialState,
 	reducers: {
 		fetchedUsers(state) {